This commit updates the following:
1- cuore build script can now deep compile test projects, by deep
compilation I mean that it first compiles and packages the main project
then compiles itself.

2- Kara-typer phase now works, it correctly types a class (It should
work for all possible cases)

3- Changed the way that transform method works in
TransormerPluginComponent class, now it returns a tuple of Tree and
Boolean, the second part of the tuple is to indicate that there is no
need to go deeper the tree. I know this should become easier and neater.

4- kara-mover phase is not done yet, but there is some basic block for
it, which does nothing but iterates through the tree.
k
kkkkkkkkj

kjj;jkl;jlkj

\

kk

:q