This commit makes some changes to kara-mover phase.
The kara-mover phase is now supposed to be working (it has not been
thoroughly tested, though). But it is far from being clean and elegant;
there are a lot of blocks of it that should go to Lombrello framework,
or even parts that are really badly engineered.

But after all, with this commit my goal was nothing just to have
something working, not something that is purely correct.