The Kara plugin is now fully working.
The implementation of Kara plugin is finished, it has everything that it
was supposed to do, except two limitation which are described in the
TODO of the source code of the ReadDependentExctractor.scala.

And the limitation is something that I suppose is not fixable (about
inner defs and types).

I also updated the simpe-plugin to conform with the latest changes of
the Lombrello framework